requirements.txt
# ============================================================
# requirements.txt — Dependencias del proyecto
# ============================================================
# Instálalas con: pip install -r requirements.txt
# ============================================================

discord.py>=2.3.0        # Librería principal para bots de Discord
aiosqlite>=0.19.0        # SQLite asíncrono (compatible con discord.py)
python-dotenv>=1.0.0     # Para leer el archivo .env con el token
flask>=3.0.0             # Panel web de configuración (dashboard.py)

httpx>=0.27.0             # HTTP async para APIs de crypto
solders>=0.21.0           # Generación de wallets Solana
coincurve>=18.0.0         # ECDSA secp256k1 para sweep LTC
